# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WatchStore Api List",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/products_by_category/{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by category id",
                "operationId": "405d1e219fade8f19f916b7aecf56cc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/products_by_brand/{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by category id",
                "operationId": "47edbc06e30ed03b55a05d3afc326cc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/product_details/{id}": {
            "get": {
                "tags": [
                    "Product Details"
                ],
                "description": "get product details data by product id",
                "operationId": "e2bda0a2b3a40d81d4f4c4d89dcbe136",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}